Tweaks to test-java-files.sh

+++ test-java-files.sh
@@ -1,20 +1,29 @@
 #!/usr/bin/env bash
 
 set -euo pipefail
-#set -x
 
-JAR=~/hacks/bhs-cs/java/target/bhs-cs.jar
-#JAR=~/hacks/jobe/jars/bhs-cs.jar
-RUNNER=com.gigamonkeys.bhs.testing.TestRunner
+BHS_CS=/Users/peter/hacks/bhs-cs
+JAR=$BHS_CS/java/target/bhs-cs.jar
+
+export BHS_DATA_ROOT=$BHS_CS/java/
+
+if java --enable-preview -Xrs -Xss8m -Xmx200m -cp "$JAR" "$RUNNER" "$tester" "$f" > "$(dirname "$f")/results.json " ; then
+
 
 tester="com.gigamonkeys.bhs.assignments.$1"
 shift
+
+if [[ "$#" -eq 0 ]]; then
+    echo "No files to test. Did you forget the tester class? (You gave $tester)"
+    exit 1
+fi
 
 for f; do
     echo -n "Testing $f ... "
     rm -f "$(dirname "$f")/broken.txt"
     if [[ ! -e "$(dirname "$f")/missing.txt" ]]; then
-        if java -Xrs -Xss8m -Xmx200m -cp "$JAR" "$RUNNER" "$tester" "$f" > "$(dirname "$f")/results.json" 2> /dev/null; then
+        if java --enable-preview -Xrs -Xss8m -Xmx200m -cp "$JAR" "$RUNNER" "$tester" "$f" > "$(dirname "$f")/results.json" ; then
+        #if java -Xrs -Xss8m -Xmx200m -cp "$JAR" "$RUNNER" "$tester" "$f" > "$(dirname "$f")/results.json" 2> /dev/null; then
             echo "ok."
         else
             echo "broken."
